Moved activity dialogs to common
