+++ client/components/repository/common/Sidebar/Discussion/TextEditor.vue
