+++ client/components/repository/common/Sidebar/Header.vue
@@ -28,7 +28,7 @@
 
 <script>
 import ActivityOptions from '@/components/repository/common/ActivityOptions';
-import CreateDialog from '@/components/repository/Outline/InsertActivity/CreateDialog';
+import CreateDialog from '@/components/repository/common/CreateDialog';
 import get from 'lodash/get';
 import { isEditable } from 'shared/activities';
 import { mapGetters } from 'vuex';
