+++ client/components/repository/common/Sidebar/Publishing.vue
