+++ client/components/repository/common/Sidebar/index.vue
